Move IPInt slow paths to a new file https://bugs.webkit.org/show_bug.cgi?id=261963
Reviewed by Yusuke Suzuki.

This is a simple code cleanliness improvement. No code should be changed
with the exception that `doWasmCallIndirect` used the wrong CALLEE macro
for IPInt before.

* Source/JavaScriptCore/JavaScriptCore.xcodeproj/project.pbxproj:
* Source/JavaScriptCore/Sources.txt:
* Source/JavaScriptCore/wasm/WasmIPIntSlowPaths.cpp: Added.
(JSC::LLInt::shouldJIT):
(JSC::LLInt::jitCompileAndSetHeuristics):
(JSC::LLInt::WASM_IPINT_EXTERN_CPP_DECL):
(JSC::LLInt::WASM_IPINT_EXTERN_CPP_DECL_1P):
(JSC::LLInt::doWasmCall):
(JSC::LLInt::doWasmCallIndirect):
* Source/JavaScriptCore/wasm/WasmIPIntSlowPaths.h: Copied from Source/JavaScriptCore/wasm/WasmSlowPaths.h.
* Source/JavaScriptCore/wasm/WasmSlowPaths.cpp:
(JSC::LLInt::WASM_SLOW_PATH_DECL):
(JSC::LLInt::WASM_IPINT_EXTERN_CPP_DECL): Deleted.
(JSC::LLInt::WASM_IPINT_EXTERN_CPP_DECL_1P): Deleted.
* Source/JavaScriptCore/wasm/WasmSlowPaths.h:

Canonical link: https://commits.webkit.org/268516@main
