+++ Source/JavaScriptCore/wasm/WasmSlowPaths.cpp
@@ -79,8 +79,6 @@ namespace JSC { namespace LLInt {
 
 #define CALLEE() \
     static_cast<Wasm::LLIntCallee*>(callFrame->callee().asNativeCallee())
-#define IPINT_CALLEE() \
-    static_cast<Wasm::IPIntCallee*>(callFrame->callee().asNativeCallee())
 
 #define READ(virtualRegister) \
     (virtualRegister.isConstant() \
@@ -98,7 +96,6 @@ extern "C" void wasm_log_crash(CallFrame*, Wasm::Instance* instance)
     RELEASE_ASSERT_NOT_REACHED();
 }
 
-// LLInt overload
 inline bool shouldJIT(Wasm::LLIntCallee* callee, RequiredWasmJIT requiredJIT = RequiredWasmJIT::Any)
 {
     if (requiredJIT == RequiredWasmJIT::OMG) {
@@ -117,26 +114,6 @@ inline bool shouldJIT(Wasm::LLIntCallee* callee, RequiredWasmJIT requiredJIT = R
     return true;
 }
 
-// IPInt overload
-inline bool shouldJIT(Wasm::IPIntCallee* callee, RequiredWasmJIT requiredJIT = RequiredWasmJIT::Any)
-{
-    if (requiredJIT == RequiredWasmJIT::OMG) {
-        if (!Options::useOMGJIT() || !Wasm::OMGPlan::ensureGlobalOMGAllowlist().containsWasmFunction(callee->functionIndex()))
-            return false;
-    } else {
-        if (Options::wasmIPIntTiersUpToBBQ()
-            && (!Options::useBBQJIT() || !Wasm::BBQPlan::ensureGlobalBBQAllowlist().containsWasmFunction(callee->functionIndex())))
-            return false;
-        if (!Options::wasmIPIntTiersUpToOMG()
-            && (!Options::useOMGJIT() || !Wasm::OMGPlan::ensureGlobalOMGAllowlist().containsWasmFunction(callee->functionIndex())))
-            return false;
-    }
-    if (!Options::wasmFunctionIndexRangeToCompile().isInRange(callee->functionIndex()))
-        return false;
-    return true;
-}
-
-// LLInt overload
 inline bool jitCompileAndSetHeuristics(Wasm::LLIntCallee* callee, Wasm::Instance* instance)
 {
     ASSERT(!instance->module().moduleInformation().usesSIMD(callee->functionIndex()));
@@ -173,57 +150,6 @@ inline bool jitCompileAndSetHeuristics(Wasm::LLIntCallee* callee, Wasm::Instance
         uint32_t functionIndex = callee->functionIndex();
         RefPtr<Wasm::Plan> plan;
         if (Options::wasmLLIntTiersUpToBBQ() && Wasm::BBQPlan::ensureGlobalBBQAllowlist().containsWasmFunction(functionIndex))
-            plan = adoptRef(*new Wasm::BBQPlan(instance->vm(), const_cast<Wasm::ModuleInformation&>(instance->module().moduleInformation()), functionIndex, callee->hasExceptionHandlers(), instance->calleeGroup(), Wasm::Plan::dontFinalize()));
-        else // No need to check OMG allow list: if we didn't want to compile this function, shouldJIT should have returned false.
-            plan = adoptRef(*new Wasm::OMGPlan(instance->vm(), Ref<Wasm::Module>(instance->module()), functionIndex, callee->hasExceptionHandlers(), instance->memory()->mode(), Wasm::Plan::dontFinalize()));
-
-        Wasm::ensureWorklist().enqueue(*plan);
-        if (UNLIKELY(!Options::useConcurrentJIT()))
-            plan->waitForCompletion();
-        else
-            tierUpCounter.optimizeAfterWarmUp();
-    }
-
-    return !!callee->replacement(instance->memory()->mode());
-}
-
-// IPInt overload
-inline bool jitCompileAndSetHeuristics(Wasm::IPIntCallee* callee, Wasm::Instance* instance)
-{
-    ASSERT(!instance->module().moduleInformation().usesSIMD(callee->functionIndex()));
-
-    Wasm::LLIntTierUpCounter& tierUpCounter = callee->tierUpCounter();
-    if (!tierUpCounter.checkIfOptimizationThresholdReached()) {
-        dataLogLnIf(Options::verboseOSR(), "    JIT threshold should be lifted.");
-        return false;
-    }
-
-    if (callee->replacement(instance->memory()->mode()))  {
-        dataLogLnIf(Options::verboseOSR(), "    Code was already compiled.");
-        tierUpCounter.optimizeSoon();
-        return true;
-    }
-
-    bool compile = false;
-    {
-        Locker locker { tierUpCounter.m_lock };
-        switch (tierUpCounter.m_compilationStatus) {
-        case Wasm::LLIntTierUpCounter::CompilationStatus::NotCompiled:
-            compile = true;
-            tierUpCounter.m_compilationStatus = Wasm::LLIntTierUpCounter::CompilationStatus::Compiling;
-            break;
-        case Wasm::LLIntTierUpCounter::CompilationStatus::Compiling:
-            tierUpCounter.optimizeAfterWarmUp();
-            break;
-        case Wasm::LLIntTierUpCounter::CompilationStatus::Compiled:
-            break;
-        }
-    }
-
-    if (compile) {
-        uint32_t functionIndex = callee->functionIndex();
-        RefPtr<Wasm::Plan> plan;
-        if (Options::wasmIPIntTiersUpToBBQ() && Wasm::BBQPlan::ensureGlobalBBQAllowlist().containsWasmFunction(functionIndex))
             plan = adoptRef(*new Wasm::BBQPlan(instance->vm(), const_cast<Wasm::ModuleInformation&>(instance->module().moduleInformation()), functionIndex, callee->hasExceptionHandlers(), instance->calleeGroup(), Wasm::Plan::dontFinalize()));
         else // No need to check OMG allow list: if we didn't want to compile this function, shouldJIT should have returned false.
             plan = adoptRef(*new Wasm::OMGPlan(instance->vm(), Ref<Wasm::Module>(instance->module()), functionIndex, callee->hasExceptionHandlers(), instance->memory()->mode(), Wasm::Plan::dontFinalize()));
@@ -301,25 +227,6 @@ WASM_SLOW_PATH_DECL(prologue_osr)
     if (!jitCompileAndSetHeuristics(callee, instance))
         WASM_RETURN_TWO(nullptr, nullptr);
 
-    WASM_RETURN_TWO(callee->replacement(instance->memory()->mode())->entrypoint().taggedPtr(), nullptr);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(prologue_osr, CallFrame* callFrame)
-{
-    Wasm::IPIntCallee* callee = IPINT_CALLEE();
-
-    if (!shouldJIT(callee)) {
-        callee->tierUpCounter().deferIndefinitely();
-        WASM_RETURN_TWO(nullptr, nullptr);
-    }
-
-    if (!Options::useWasmIPIntPrologueOSR())
-        WASM_RETURN_TWO(nullptr, nullptr);
-
-    dataLogLnIf(Options::verboseOSR(), *callee, ": Entered prologue_osr with tierUpCounter = ", callee->tierUpCounter());
-
-    if (!jitCompileAndSetHeuristics(callee, instance))
-        WASM_RETURN_TWO(nullptr, nullptr);
     WASM_RETURN_TWO(callee->replacement(instance->memory()->mode())->entrypoint().taggedPtr(), nullptr);
 }
 
@@ -422,114 +329,6 @@ WASM_SLOW_PATH_DECL(loop_osr)
     }
 }
 
-WASM_IPINT_EXTERN_CPP_DECL(loop_osr, CallFrame* callFrame, uint32_t pc, uint64_t *pl)
-{
-    Wasm::IPIntCallee* callee = IPINT_CALLEE();
-    Wasm::LLIntTierUpCounter& tierUpCounter = callee->tierUpCounter();
-
-    if (!Options::useWebAssemblyOSR() || !Options::useWasmIPIntLoopOSR() || !shouldJIT(callee, RequiredWasmJIT::OMG)) {
-        ipint_extern_prologue_osr(instance, callFrame);
-        WASM_RETURN_TWO(nullptr, nullptr);
-    }
-
-    dataLogLnIf(Options::verboseOSR(), *callee, ": Entered loop_osr with tierUpCounter = ", callee->tierUpCounter());
-
-    if (!tierUpCounter.checkIfOptimizationThresholdReached()) {
-        dataLogLnIf(Options::verboseOSR(), "    JIT threshold should be lifted.");
-        WASM_RETURN_TWO(nullptr, nullptr);
-    }
-
-    unsigned loopOSREntryBytecodeOffset = pc;
-    const auto& osrEntryData = tierUpCounter.osrEntryDataForLoop(loopOSREntryBytecodeOffset);
-
-    if (Options::wasmIPIntTiersUpToBBQ() && Options::useBBQJIT()) {
-        if (!jitCompileAndSetHeuristics(callee, instance))
-            WASM_RETURN_TWO(nullptr, nullptr);
-
-        Wasm::BBQCallee* bbqCallee;
-        {
-            Locker locker { instance->calleeGroup()->m_lock };
-            bbqCallee = instance->calleeGroup()->bbqCallee(locker, callee->functionIndex());
-        }
-        RELEASE_ASSERT(bbqCallee);
-
-        size_t osrEntryScratchBufferSize = bbqCallee->osrEntryScratchBufferSize();
-        RELEASE_ASSERT(osrEntryScratchBufferSize >= osrEntryData.values.size());
-
-        uint64_t* buffer = instance->vm().wasmContext.scratchBufferForSize(osrEntryScratchBufferSize);
-        if (!buffer)
-            WASM_RETURN_TWO(nullptr, nullptr);
-
-        uint32_t index = 0;
-        buffer[index++] = osrEntryData.loopIndex;
-        for (uint32_t i = 0; i < callee->m_numLocals; ++i)
-            buffer[index++] = pl[i];
-        while (index < osrEntryData.values.size()) {
-            pl -= 2; // each stack slot is 16B
-            buffer[index++] = *pl;
-        }
-
-        auto sharedLoopEntrypoint = bbqCallee->sharedLoopEntrypoint();
-        RELEASE_ASSERT(sharedLoopEntrypoint);
-        WASM_RETURN_TWO(buffer, sharedLoopEntrypoint->taggedPtr());
-    } else {
-        const auto doOSREntry = [&](Wasm::OSREntryCallee* osrEntryCallee) {
-            if (osrEntryCallee->loopIndex() != osrEntryData.loopIndex)
-                WASM_RETURN_TWO(nullptr, nullptr);
-
-            size_t osrEntryScratchBufferSize = osrEntryCallee->osrEntryScratchBufferSize();
-            RELEASE_ASSERT(osrEntryScratchBufferSize == osrEntryData.values.size());
-
-            uint64_t* buffer = instance->vm().wasmContext.scratchBufferForSize(osrEntryScratchBufferSize);
-            if (!buffer)
-                WASM_RETURN_TWO(nullptr, nullptr);
-
-            uint32_t index = 0;
-            for (uint32_t i = 0; i < callee->m_numLocals; ++i)
-                buffer[index++] = pl[i];
-            while (index < osrEntryData.values.size()) {
-                pl -= 2; // each stack slot is 16B
-                buffer[index++] = *pl;
-            }
-
-            WASM_RETURN_TWO(buffer, osrEntryCallee->entrypoint().taggedPtr());
-        };
-
-        if (auto* osrEntryCallee = callee->osrEntryCallee(instance->memory()->mode()))
-            return doOSREntry(osrEntryCallee);
-
-        bool compile = false;
-        {
-            Locker locker { tierUpCounter.m_lock };
-            switch (tierUpCounter.m_loopCompilationStatus) {
-            case Wasm::LLIntTierUpCounter::CompilationStatus::NotCompiled:
-                compile = true;
-                tierUpCounter.m_loopCompilationStatus = Wasm::LLIntTierUpCounter::CompilationStatus::Compiling;
-                break;
-            case Wasm::LLIntTierUpCounter::CompilationStatus::Compiling:
-                tierUpCounter.optimizeAfterWarmUp();
-                break;
-            case Wasm::LLIntTierUpCounter::CompilationStatus::Compiled:
-                break;
-            }
-        }
-
-        if (compile) {
-            Ref<Wasm::Plan> plan = adoptRef(*static_cast<Wasm::Plan*>(new Wasm::OSREntryPlan(instance->vm(), Ref<Wasm::Module>(instance->module()), Ref<Wasm::Callee>(*callee), callee->functionIndex(), callee->hasExceptionHandlers(), osrEntryData.loopIndex, instance->memory()->mode(), Wasm::Plan::dontFinalize())));
-            Wasm::ensureWorklist().enqueue(plan.copyRef());
-            if (UNLIKELY(!Options::useConcurrentJIT()))
-                plan->waitForCompletion();
-            else
-                tierUpCounter.optimizeAfterWarmUp();
-        }
-
-        if (auto* osrEntryCallee = callee->osrEntryCallee(instance->memory()->mode()))
-            return doOSREntry(osrEntryCallee);
-
-        WASM_RETURN_TWO(nullptr, nullptr);
-    }
-}
-
 WASM_SLOW_PATH_DECL(epilogue_osr)
 {
     Wasm::LLIntCallee* callee = CALLEE();
@@ -545,23 +344,6 @@ WASM_SLOW_PATH_DECL(epilogue_osr)
 
     jitCompileAndSetHeuristics(callee, instance);
     WASM_END_IMPL();
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(epilogue_osr, CallFrame* callFrame)
-{
-    Wasm::IPIntCallee* callee = IPINT_CALLEE();
-
-    if (!shouldJIT(callee)) {
-        callee->tierUpCounter().deferIndefinitely();
-        WASM_RETURN_TWO(nullptr, nullptr);
-    }
-    if (!Options::useWasmIPIntEpilogueOSR())
-        WASM_RETURN_TWO(nullptr, nullptr);
-
-    dataLogLnIf(Options::verboseOSR(), *callee, ": Entered epilogue_osr with tierUpCounter = ", callee->tierUpCounter());
-
-    jitCompileAndSetHeuristics(callee, instance);
-    WASM_RETURN_TWO(nullptr, nullptr);
 }
 
 WASM_SLOW_PATH_DECL(simd_go_straight_to_bbq_osr)
@@ -619,17 +401,6 @@ WASM_SLOW_PATH_DECL(ref_func)
 {
     auto instruction = pc->as<WasmRefFunc>();
     WASM_RETURN(Wasm::refFunc(instance, instruction.m_functionIndex));
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(ref_func, unsigned index)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    WASM_RETURN_TWO(bitwise_cast<void*>(Wasm::refFunc(instance, index)), 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(index);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
 }
 
 WASM_SLOW_PATH_DECL(array_new)
@@ -757,21 +528,6 @@ WASM_SLOW_PATH_DECL(table_get)
     WASM_RETURN(result);
 }
 
-WASM_IPINT_EXTERN_CPP_DECL(table_get, unsigned tableIndex, unsigned index)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    EncodedJSValue result = Wasm::tableGet(instance, tableIndex, index);
-    if (!result)
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<uint64_t>(Wasm::ExceptionType::OutOfBoundsTableAccess)));
-    WASM_RETURN_TWO(bitwise_cast<void*>(result), 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(tableIndex);
-    UNUSED_PARAM(index);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARMv8 and X86_64 (for now)");
-#endif
-}
-
 WASM_SLOW_PATH_DECL(table_set)
 {
     auto instruction = pc->as<WasmTableSet>();
@@ -780,21 +536,6 @@ WASM_SLOW_PATH_DECL(table_set)
     if (!Wasm::tableSet(instance, instruction.m_tableIndex, index, value))
         WASM_THROW(Wasm::ExceptionType::OutOfBoundsTableAccess);
     WASM_END();
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(table_set, unsigned tableIndex, unsigned index, EncodedJSValue value)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    if (!Wasm::tableSet(instance, tableIndex, index, value))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<uint64_t>(Wasm::ExceptionType::OutOfBoundsTableAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(tableIndex);
-    UNUSED_PARAM(index);
-    UNUSED_PARAM(value);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
 }
 
 WASM_SLOW_PATH_DECL(table_init)
@@ -808,24 +549,6 @@ WASM_SLOW_PATH_DECL(table_init)
     WASM_END();
 }
 
-WASM_IPINT_EXTERN_CPP_DECL(table_init, uint32_t* metadata, uint32_t dest, uint64_t srcAndLength)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    uint32_t dstOffset = dest;
-    uint32_t srcOffset = srcAndLength >> 32;
-    uint32_t length = srcAndLength & 0xffffffff;
-    if (!Wasm::tableInit(instance, metadata[0], metadata[1], dstOffset, srcOffset, length))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<uint64_t>(Wasm::ExceptionType::OutOfBoundsTableAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(metadata);
-    UNUSED_PARAM(dest);
-    UNUSED_PARAM(srcAndLength);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
-}
-
 WASM_SLOW_PATH_DECL(table_fill)
 {
     auto instruction = pc->as<WasmTableFill>();
@@ -837,53 +560,12 @@ WASM_SLOW_PATH_DECL(table_fill)
     WASM_END();
 }
 
-WASM_IPINT_EXTERN_CPP_DECL(table_fill, uint32_t tableIndex, EncodedJSValue fill, int64_t offsetAndSize)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    uint32_t offset = offsetAndSize >> 32;
-    uint32_t size = offsetAndSize & 0xffffffff;
-    if (!Wasm::tableFill(instance, tableIndex, offset, fill, size))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<uint64_t>(Wasm::ExceptionType::OutOfBoundsTableAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(tableIndex);
-    UNUSED_PARAM(fill);
-    UNUSED_PARAM(offsetAndSize);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
-}
-
 WASM_SLOW_PATH_DECL(table_grow)
 {
     auto instruction = pc->as<WasmTableGrow>();
     EncodedJSValue fill = READ(instruction.m_fill).encodedJSValue();
     uint32_t size = READ(instruction.m_size).unboxedUInt32();
     WASM_RETURN(Wasm::tableGrow(instance, instruction.m_tableIndex, fill, size));
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(table_grow, int32_t tableIndex, EncodedJSValue fill, uint32_t size)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    WASM_RETURN_TWO(bitwise_cast<void*>(Wasm::tableGrow(instance, tableIndex, fill, size)), 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(tableIndex);
-    UNUSED_PARAM(fill);
-    UNUSED_PARAM(size);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
-}
-
-WASM_IPINT_EXTERN_CPP_DECL_1P(current_memory)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    size_t size = instance->memory()->handle().size() >> 16;
-    WASM_RETURN_TWO(bitwise_cast<void*>(size), 0);
-#else
-    UNUSED_PARAM(instance);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
 }
 
 WASM_SLOW_PATH_DECL(grow_memory)
@@ -893,95 +575,6 @@ WASM_SLOW_PATH_DECL(grow_memory)
     auto instruction = pc->as<WasmGrowMemory>();
     int32_t delta = READ(instruction.m_delta).unboxedInt32();
     WASM_RETURN(Wasm::growMemory(instance, delta));
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(memory_grow, int32_t delta)
-{
-    WASM_RETURN_TWO(reinterpret_cast<void*>(Wasm::growMemory(instance, delta)), 0);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(memory_init, int32_t dataIndex, int32_t dst, int64_t srcAndLength)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    if (!Wasm::memoryInit(instance, dataIndex, dst, srcAndLength >> 32, srcAndLength & 0xffffffff))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<int64_t>(Wasm::ExceptionType::OutOfBoundsMemoryAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(dataIndex);
-    UNUSED_PARAM(dst);
-    UNUSED_PARAM(srcAndLength);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(data_drop, int32_t dataIndex)
-{
-    Wasm::dataDrop(instance, dataIndex);
-    WASM_RETURN_TWO(0, 0);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(memory_copy, int32_t dst, int32_t src, int32_t count)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    if (!Wasm::memoryCopy(instance, dst, src, count))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<int64_t>(Wasm::ExceptionType::OutOfBoundsMemoryAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(dst);
-    UNUSED_PARAM(src);
-    UNUSED_PARAM(count);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(memory_fill, int32_t dst, int32_t targetValue, int32_t count)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    if (!Wasm::memoryFill(instance, dst, targetValue, count))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<int64_t>(Wasm::ExceptionType::OutOfBoundsMemoryAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(dst);
-    UNUSED_PARAM(targetValue);
-    UNUSED_PARAM(count);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(elem_drop, int32_t dataIndex)
-{
-    Wasm::elemDrop(instance, dataIndex);
-    WASM_RETURN_TWO(0, 0);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(table_copy, int32_t* metadata, int32_t dst, int64_t srcAndCount)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    if (!Wasm::tableCopy(instance, metadata[0], metadata[1], dst, srcAndCount >> 32, srcAndCount & 0xffffffff))
-        WASM_RETURN_TWO(bitwise_cast<void*>(1L), bitwise_cast<void*>(static_cast<int64_t>(Wasm::ExceptionType::OutOfBoundsTableAccess)));
-    WASM_RETURN_TWO(0, 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(metadata);
-    UNUSED_PARAM(dst);
-    UNUSED_PARAM(srcAndCount);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(table_size, int32_t tableIndex)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    int32_t result = Wasm::tableSize(instance, tableIndex);
-    WASM_RETURN_TWO(bitwise_cast<void*>(static_cast<EncodedJSValue>(result)), 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(tableIndex);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supported on ARM64 and X86_64 (for now)");
-#endif
 }
 
 inline UGPRPair doWasmCall(Wasm::Instance* instance, unsigned functionIndex)
@@ -999,11 +592,6 @@ inline UGPRPair doWasmCall(Wasm::Instance* instance, unsigned functionIndex)
     }
 
     WASM_CALL_RETURN(instance, codePtr.taggedPtr(), WasmEntryPtrTag);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(call, unsigned functionIndex)
-{
-    return doWasmCall(instance, functionIndex);
 }
 
 WASM_SLOW_PATH_DECL(call)
@@ -1027,27 +615,6 @@ inline UGPRPair doWasmCallIndirect(CallFrame* callFrame, Wasm::Instance* instanc
         WASM_THROW(Wasm::ExceptionType::NullTableEntry);
 
     const auto& callSignature = CALLEE()->signature(typeIndex);
-    if (callSignature.index() != function.m_function.typeIndex)
-        WASM_THROW(Wasm::ExceptionType::BadSignature);
-
-    WASM_CALL_RETURN(function.m_instance, function.m_function.entrypointLoadLocation->taggedPtr(), WasmEntryPtrTag);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(call_indirect, CallFrame* callFrame, unsigned functionIndex, unsigned* metadataEntry)
-{
-    unsigned tableIndex = metadataEntry[0];
-    unsigned typeIndex = metadataEntry[1];
-    Wasm::FuncRefTable* table = instance->table(tableIndex)->asFuncrefTable();
-
-    if (functionIndex >= table->length())
-        WASM_THROW(Wasm::ExceptionType::OutOfBoundsCallIndirect);
-
-    const Wasm::FuncRefTable::Function& function = table->function(functionIndex);
-
-    if (function.m_function.typeIndex == Wasm::TypeDefinition::invalidIndex)
-        WASM_THROW(Wasm::ExceptionType::NullTableEntry);
-
-    const auto& callSignature = static_cast<Wasm::IPIntCallee*>(callFrame->callee().asNativeCallee())->signature(typeIndex);
     if (callSignature.index() != function.m_function.typeIndex)
         WASM_THROW(Wasm::ExceptionType::BadSignature);
 
@@ -1249,34 +816,11 @@ WASM_SLOW_PATH_DECL(call_builtin)
     WASM_END();
 }
 
-WASM_IPINT_EXTERN_CPP_DECL(set_global_ref, uint32_t globalIndex, JSValue value)
-{
-    instance->setGlobal(globalIndex, value);
-    WASM_RETURN_TWO(0, 0);
-}
-
 WASM_SLOW_PATH_DECL(set_global_ref)
 {
     auto instruction = pc->as<WasmSetGlobalRef>();
     instance->setGlobal(instruction.m_globalIndex, READ(instruction.m_value).jsValue());
     WASM_END_IMPL();
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(get_global_64, unsigned index)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    WASM_RETURN_TWO(bitwise_cast<void*>(instance->loadI64Global(index)), 0);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(index);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(set_global_64, unsigned index, uint64_t value)
-{
-    instance->setGlobal(index, value);
-    WASM_RETURN_TWO(0, 0);
 }
 
 WASM_SLOW_PATH_DECL(set_global_ref_portable_binding)
@@ -1299,21 +843,6 @@ WASM_SLOW_PATH_DECL(memory_atomic_wait32)
     WASM_RETURN(result);
 }
 
-
-WASM_IPINT_EXTERN_CPP_DECL(memory_atomic_wait32, uint64_t pointerWithOffset, uint32_t value, uint64_t timeout)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    int32_t result = Wasm::memoryAtomicWait32(instance, pointerWithOffset, value, timeout);
-    WASM_RETURN_TWO(bitwise_cast<void*>(static_cast<intptr_t>(result)), nullptr);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(pointerWithOffset);
-    UNUSED_PARAM(value);
-    UNUSED_PARAM(timeout);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
-}
-
 WASM_SLOW_PATH_DECL(memory_atomic_wait64)
 {
     auto instruction = pc->as<WasmMemoryAtomicWait64>();
@@ -1327,20 +856,6 @@ WASM_SLOW_PATH_DECL(memory_atomic_wait64)
     WASM_RETURN(result);
 }
 
-WASM_IPINT_EXTERN_CPP_DECL(memory_atomic_wait64, uint64_t pointerWithOffset, uint64_t value, uint64_t timeout)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    int32_t result = Wasm::memoryAtomicWait64(instance, pointerWithOffset, value, timeout);
-    WASM_RETURN_TWO(bitwise_cast<void*>(static_cast<intptr_t>(result)), nullptr);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(pointerWithOffset);
-    UNUSED_PARAM(value);
-    UNUSED_PARAM(timeout);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
-}
-
 WASM_SLOW_PATH_DECL(memory_atomic_notify)
 {
     auto instruction = pc->as<WasmMemoryAtomicNotify>();
@@ -1351,20 +866,6 @@ WASM_SLOW_PATH_DECL(memory_atomic_notify)
     if (result < 0)
         WASM_THROW(Wasm::ExceptionType::OutOfBoundsMemoryAccess);
     WASM_RETURN(result);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(memory_atomic_notify, unsigned base, unsigned offset, int32_t count)
-{
-#if CPU(ARM64) || CPU(X86_64)
-    int32_t result = Wasm::memoryAtomicNotify(instance, base, offset, count);
-    WASM_RETURN_TWO(bitwise_cast<void*>(static_cast<intptr_t>(result)), nullptr);
-#else
-    UNUSED_PARAM(instance);
-    UNUSED_PARAM(base);
-    UNUSED_PARAM(offset);
-    UNUSED_PARAM(count);
-    RELEASE_ASSERT_NOT_REACHED("IPInt only supports ARM64 and X86_64 (for now)");
-#endif
 }
 
 WASM_SLOW_PATH_DECL(throw)
@@ -1382,26 +883,6 @@ WASM_SLOW_PATH_DECL(throw)
     for (unsigned i = 0; i < tag.parameterCount(); ++i)
         values[i] = READ((instruction.m_firstValue - i)).encodedJSValue();
 
-    JSWebAssemblyException* exception = JSWebAssemblyException::create(vm, globalObject->webAssemblyExceptionStructure(), tag, WTFMove(values));
-    throwException(globalObject, throwScope, exception);
-
-    genericUnwind(vm, callFrame);
-    ASSERT(!!vm.callFrameForCatch);
-    ASSERT(!!vm.targetMachinePCForThrow);
-    WASM_RETURN_TWO(vm.targetMachinePCForThrow, nullptr);
-}
-
-WASM_IPINT_EXTERN_CPP_DECL(throw, CallFrame* callFrame, uint32_t exceptionIndex)
-{
-    SlowPathFrameTracer tracer(instance->vm(), callFrame);
-
-    JSGlobalObject* globalObject = instance->globalObject();
-    VM& vm = globalObject->vm();
-    auto throwScope = DECLARE_THROW_SCOPE(vm);
-
-    const Wasm::Tag& tag = instance->tag(exceptionIndex);
-
-    FixedVector<uint64_t> values(tag.parameterCount());
     JSWebAssemblyException* exception = JSWebAssemblyException::create(vm, globalObject->webAssemblyExceptionStructure(), tag, WTFMove(values));
     throwException(globalObject, throwScope, exception);
